# docs/designs/api/US-2.4-openapi.yaml
openapi: 3.0.3
info:
  title: Password Reset (US-2.4 / spec US-2.4) — fragment
  version: "1.0"
paths:
  /v1/auth/password-reset/request:
    post:
      operationId: requestPasswordReset
      summary: Issue a single-use password-reset token and email it to the account
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetRequestRequest"
      responses:
        "202":
          description: >
            Always returned with the same generic body regardless of
            whether the email resolves to a registered, active account
            (FR-1), an unknown address, or a deactivated account (FR-3) —
            NFR-002 anti-enumeration. On a real, eligible account: a
            32-byte `secrets.token_urlsafe(32)` reset token is generated
            (SHA-256 hash persisted only), any prior unconsumed token for
            the account is invalidated, an email is fired containing the
            token in the URL fragment (never the query string, per
            NFR-001), and an `auth_audit_log` entry
            (`event=password_reset_requested`) is written. For an unknown
            or deactivated account, no token is created and no email is
            sent, but the audit entry is still written (OD-3) — a
            server-side-only fact that does not alter this response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasswordResetRequestResponse"
        "429":
          description: >
            One of three limits tripped on this account/IP, checked in
            order (resolved OD-2): a 60-second per-account cooldown since
            the last request, then a 5-requests/hour per-account limit,
            then a 10-requests/hour per-IP limit. The response carries the
            `Retry-After` of whichever limit was the first to trip.
          headers:
            Retry-After:
              schema:
                type: integer
                description: Seconds until the tripped limit's window clears.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/TooManyAttemptsProblem"
  /v1/auth/password-reset/confirm:
    post:
      operationId: confirmPasswordReset
      summary: Consume a reset token and set a new password
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetConfirmRequest"
      responses:
        "200":
          description: >
            Token was valid, unconsumed, unexpired, and the new password
            passed policy (FR-2). Consumption is atomic
            (`UPDATE...WHERE consumed_at IS NULL RETURNING`, spec-review
            resolution accepted 2026-09-01) so two concurrent confirm
            calls against the same token can never both succeed. The
            password hash is replaced (Argon2id); `revoke_before:{user_id}`
            is set to now, terminating every existing session and refresh
            family; a "your password was changed" notification is sent;
            an `auth_audit_log` entry (`event=password_reset_completed`)
            is written. Empty body — no token is issued here, matching the
            source story's Success column (`200`, no response schema
            stated) and the sibling US-2.1/US-2.2 pattern of not
            auto-logging-in after a security-sensitive credential change.
        "400":
          description: >
            Token is expired, already consumed (including the losing side
            of the atomic race above), or matches no stored hash (FR-4).
            An unknown hash and an already-consumed token both respond
            with `token-invalid`; an expired token responds with
            `token-expired` — mapping resolved by precedent
            (`email_verification` module), since the source story left it
            open. The existing password is unchanged; the body signals the
            option to request a new link via the `type` slug alone (the
            source story states no separate response field for this).
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/TokenExpiredProblem"
                  - $ref: "#/components/schemas/TokenInvalidProblem"
        "422":
          description: >
            New password is shorter than 12 characters, appears in the
            local breached-password list/bloom filter (resolved OD-1 — no
            live network call), or equals the account's current password
            (FR-5). The token is NOT consumed by this rejection, so the
            same link can be retried. `errors` states which rule(s)
            failed.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/PasswordPolicyProblem"
components:
  schemas:
    PasswordResetRequestRequest:
      type: object
      required: [email]
      additionalProperties: false
      properties:
        email:
          type: string
          description: >
            Plain string, no dedicated format validation — matches
            `LoginRequest.email` precedent (US-2.1/US-2.1); this codebase
            validates no endpoint's email field at the schema layer.
    PasswordResetRequestResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
          enum: ["If an account exists, an email has been sent"]
      description: >
        Generic body per FR-1/FR-3's verbatim source text. Identical for
        every input — no field varies by account existence or status.
    PasswordResetConfirmRequest:
      type: object
      required: [token, new_password]
      additionalProperties: false
      properties:
        token:
          type: string
          description: >
            The raw (unhashed) 32-byte urlsafe token from the URL
            fragment, POSTed by the SPA — never sent as a query parameter
            (NFR-001).
        new_password:
          type: string
          format: password
          minLength: 12
          description: >
            Rejected with 422 (PasswordPolicyProblem) if shorter than 12
            chars, breached, or equal to the current password — enforced
            at the service layer (breach/equality checks require a
            database/list lookup no schema-level validator can perform),
            not solely by this `minLength` constraint.
    TooManyAttemptsProblem:
      allOf:
        - $ref: "#/components/schemas/ProblemBase"
      properties:
        type:
          type: string
          example: ".../errors/too-many-attempts"
        status:
          type: integer
          example: 429
      description: Type slug reused from login/refresh (US-2.1/US-2.3) — not a new slug introduced by this story.
    TokenExpiredProblem:
      allOf:
        - $ref: "#/components/schemas/ProblemBase"
      properties:
        type:
          type: string
          example: ".../errors/token-expired"
        status:
          type: integer
          example: 400
      description: Type slug reused from email verification (US-1.2) — not a new slug introduced by this story.
    TokenInvalidProblem:
      allOf:
        - $ref: "#/components/schemas/ProblemBase"
      properties:
        type:
          type: string
          example: ".../errors/token-invalid"
        status:
          type: integer
          example: 400
      description: Type slug reused from email verification (US-1.2) — not a new slug introduced by this story.
    PasswordPolicyProblem:
      allOf:
        - $ref: "#/components/schemas/ProblemBase"
      properties:
        type:
          type: string
          example: ".../errors/password-policy"
        status:
          type: integer
          example: 422
        errors:
          type: array
          description: One entry per failed rule (length / breached / reused), per the spec's FR-5.
          items:
            type: object
            required: [rule, message]
            properties:
              rule:
                type: string
                enum: [min_length, breached, reused]
              message:
                type: string
      description: >
        Introduces the `password-policy` type slug (per the source
        story's own Error Envelope section). Extends `ProblemBase` with an
        `errors` array — the only story in this project's Epic 2 whose
        problem body carries a field beyond the RFC 7807 base.
    ProblemBase:
      type: object
      description: RFC 7807 problem body. `type` is an absolute/relative URI ending in the slug named per response.
      required: [type, title, status, detail, instance]
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          example: /v1/auth/password-reset/confirm